requirements.txt
fastapi
uvicorn
python-dotenv
python-amazon-sp-api